switch to underscore for Meteor
Underscore is already included into Meteor core, so every Meteor app currently
has underscore included. Since lodash is pretty big this seems like the better
choice.

We will still have a weak dependency on lodash, and will use lodash if somebody
already haves it in their app. (it has to be 'stevezhu:lodash').

(lodash still is the default for node)