update required node version to 8
Node 6 is almost end-of-life, 7 already is.

I wanted to update to node 10, since 8 is also already in maintenance
LTS state, but it seems that would break @gruijter his homey app.  Since
if I'm not mistaken, Homey runs on node 8.  So node 8 it is.

This change should make the generated code more readable and shorter,
and hopefully more performant.