messages: fix when raw.Ontvangers is undefined (#122)
For some reason, not all messages sent by our school have a recipients array (don't ask me why.. makes no sense to me either). This issue mainly occurs in school-wide announcements. When fetching the message it would throw an error since the `raw.Ontvangers` object would be undefined. In order to circumvent this problem, I decided to pass in an empty array if that's the case.